+++ .eslintrc
@@ -7,6 +7,7 @@
     "ecmaVersion": 2019
   },
   "rules": {
+    "import/no-unused-modules": "off",
     "import/no-extraneous-dependencies": "off",
     "react/prop-types": "warn",
     "jsx-a11y/label-has-associated-control": "warn",
