+++ src/InQueueMessagingPlugin.js
@@ -1,8 +1,12 @@
 import { VERSION } from '@twilio/flex-ui';
 import { FlexPlugin } from 'flex-plugin';
+import PhoneCallbackIcon from '@material-ui/icons/PhoneCallback';
+import React from 'react';
+import VoicemailIcon from '@material-ui/icons/Voicemail';
 
-import { InQueueMessagingHelper, logger } from './helpers';
+import { logger } from './helpers';
 import reducers, { namespace } from './states';
+import { CallbackComponent, VoicemailComponent } from './components';
 
 const PLUGIN_NAME = 'InQueueMessagingPlugin';
 
@@ -21,8 +25,65 @@ export default class InQueueMessagingPlugin extends FlexPlugin {
   async init(flex, manager) {
     this.registerReducers(manager);
 
-    //  merge in callback/voice mail JSX
-    InQueueMessagingHelper(flex, manager);
+    this.registerCallbackChannel(flex, manager);
+    this.registerVoicemailChannel(flex, manager);
+  }
+
+  /**
+   * Registers the {@link CallbackComponent}
+   */
+  registerCallbackChannel(flex, manager) {
+    // Create Voicemail Channel
+    const CallbackChannel = flex.DefaultTaskChannels.createDefaultTaskChannel(
+      'callback',
+      (task) => task.taskChannelUniqueName === 'callback',
+      'CallbackIcon',
+      'CallbackIcon',
+      'palegreen',
+    );
+    // Basic Voicemail Channel Settings
+    CallbackChannel.templates.TaskListItem.firstLine = (task) => `${task.queueName}: ${task.attributes.name}`;
+    CallbackChannel.templates.TaskCanvasHeader.title = (task) => `${task.queueName}: ${task.attributes.name}`;
+    CallbackChannel.templates.IncomingTaskCanvas.firstLine = (task) => task.queueName;
+    // Lead Channel Icon
+    CallbackChannel.icons.active = <PhoneCallbackIcon key="active-callback-icon" />;
+    CallbackChannel.icons.list = <PhoneCallbackIcon key="list-callback-icon" />;
+    CallbackChannel.icons.main = <PhoneCallbackIcon key="main-callback-icon" />;
+    // Register Lead Channel
+    flex.TaskChannels.register(CallbackChannel);
+
+    flex.TaskInfoPanel.Content.replace(<CallbackComponent key="demo-component" manager={manager} />, {
+      sortOrder: -1,
+      if: (props) => props.task.attributes.taskType === 'callback',
+    });
+  }
+
+  /**
+   * Registers the {@link VoicemailComponent}
+   */
+  registerVoicemailChannel(flex, manager) {
+    const VoiceMailChannel = flex.DefaultTaskChannels.createDefaultTaskChannel(
+      'voicemail',
+      (task) => task.taskChannelUniqueName === 'voicemail',
+      'VoicemailIcon',
+      'VoicemailIcon',
+      'deepskyblue',
+    );
+    // Basic Voicemail Channel Settings
+    VoiceMailChannel.templates.TaskListItem.firstLine = (task) => `${task.queueName}: ${task.attributes.name}`;
+    VoiceMailChannel.templates.TaskCanvasHeader.title = (task) => `${task.queueName}: ${task.attributes.name}`;
+    VoiceMailChannel.templates.IncomingTaskCanvas.firstLine = (task) => task.queueName;
+    // Lead Channel Icon
+    VoiceMailChannel.icons.active = <VoicemailIcon key="active-voicemail-icon" />;
+    VoiceMailChannel.icons.list = <VoicemailIcon key="list-voicemail-icon" />;
+    VoiceMailChannel.icons.main = <VoicemailIcon key="main-voicemail-icon" />;
+    // Register Lead Channel
+    flex.TaskChannels.register(VoiceMailChannel);
+
+    flex.TaskInfoPanel.Content.replace(<VoicemailComponent key="demo-component" manager={manager} />, {
+      sortOrder: -1,
+      if: (props) => props.task.attributes.taskType === 'voicemail',
+    });
   }
 
   /**
