+++ src/components/callback/CallbackStyles.js
@@ -1,0 +1,25 @@
+export default {
+  itemWrapper: {
+    width: '100%',
+  },
+  itemBold: { fontWeight: 'bold' },
+  item: {
+    width: 100,
+  },
+  itemDetail: {
+    textAlign: 'right',
+    float: 'right',
+    marginRight: '5px',
+    marginTop: '3px',
+  },
+  cbButton: {
+    width: '100%',
+    marginBottom: '5px',
+    fontSize: '10pt',
+  },
+  textCenter: {
+    textAlign: 'center',
+    color: 'blue',
+  },
+  info: { position: 'relative', top: '3px' },
+};
