+++ src/components/callback/index.js
@@ -1,0 +1,3 @@
+import CallbackContainer from './CallbackContainer';
+
+export default CallbackContainer;
