+++ src/components/common/index.js
@@ -1,0 +1,3 @@
+import * as inqueueUtils from './inqueueUtils';
+
+export { inqueueUtils };
