+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as CallbackComponent } from './callback';
+export { default as VoicemailComponent } from './voicemail';
