+++ src/components/voicemail/VoicemailComponent.js
@@ -2,79 +2,16 @@ import React from 'react';
 import * as Flex from '@twilio/flex-ui';
 import moment from 'moment';
 import 'moment-timezone';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
 import Button from '@material-ui/core/Button';
 import TextField from '@material-ui/core/TextField';
 import Tooltip from '@material-ui/core/Tooltip';
 import Icon from '@material-ui/core/Icon';
 
-import { buildUrl, logger, http } from '../../helpers';
-import { Actions } from '../../states/ActionInQueueMessagingState';
+import styles from './VoicemailStyles';
+import { inqueueUtils } from '../common';
 
-const styles = {
-  itemWrapper: {
-    width: '100%',
-  },
-  itemBold: { fontWeight: 'bold' },
-  item: {
-    width: 150,
-  },
-  itemDetail: {
-    textAlign: 'right',
-    float: 'right',
-    marginRight: '5px',
-    marginTop: '3px',
-  },
-  cbButton: {
-    width: '100%',
-    marginBottom: '5px',
-    fontSize: '9pt',
-  },
-  audioWrapper: {
-    width: '100%',
-  },
-  transcriptWrapper: {
-    width: '100%',
-    marginBottom: 10,
-  },
-  h4Title: {
-    fontWeight: 'bold',
-  },
-  transcript: {
-    width: '90%',
-    marginRight: 10,
-    height: 50,
-    paddingLeft: 10,
-    paddingRight: 10,
-    paddingTop: 10,
-    paddingBottom: 10,
-    border: 'solid 1px #999999',
-  },
-  audio: {
-    width: '100%',
-    marginTop: 10,
-    marginBottom: 10,
-  },
-  textCenter: {
-    textAlign: 'center',
-    color: 'blue',
-  },
-  textAlert: {
-    color: 'red',
-    textAlign: 'center',
-  },
-  test: {
-    fontSize: '9pt',
-  },
-  info: { position: 'relative', top: '3px' },
-};
-
-const inqueueUtilsUri = '/inqueue-utils';
-// const [checked, setChecked] = React.useState(true);
-
-class InQueueVoicemailComponent extends React.Component {
-  static displayName = 'InQueueVoicemailComponent';
+export default class VoicemailComponent extends React.Component {
+  static displayName = 'VoicemailComponent';
 
   constructor(props) {
     super(props);
@@ -86,88 +23,16 @@ class InQueueVoicemailComponent extends React.Component {
    * create outbound call from Flex using Actions API 'StartOutboundCall'
    *
    */
-  vmCallButtonAccessibility = async (state) => {
-    const mgr = Flex.Manager.getInstance();
-    const { attributes, taskSid } = this.props.task;
-    const data = {
-      mode: 'UiPlugin',
-      type: 'voicemail',
-      Token: mgr.user.token,
-      taskSid,
-      attributes,
-      state,
-    };
+  vmCallButtonAccessibility = async (state) => inqueueUtils(this.props.task, 'voicemail', state);
 
-    return http
-      .post(buildUrl(inqueueUtilsUri), data, { noJson: true })
-      .then(() => {
-        logger.debug('==== cbUiPlugin web service success ====');
-      })
-      .catch((error) => {
-        logger.error('cbUiPlugin web service error', error);
-      });
-  };
-
-  startTransfer = async () => {
-    const mgr = Flex.Manager.getInstance();
-    const { attributes, taskSid, workflowSid, queueName } = this.props.task;
-
-    const data = {
-      mode: 'requeueTasks',
-      type: 'voicemail',
-      Token: mgr.user.token,
-      taskSid,
-      attributes,
-      workflowSid,
-      queueName,
-      state: false,
-    };
-
-    return http
-      .post(buildUrl(inqueueUtilsUri), data)
-      .then(() => {
-        logger.debug('==== requeue web service success ====');
-
-        /*
-         *   enable calling on next retry
-         * this.props.vmCallButtonDisable(false);
-         */
-      })
-      .catch((error) => {
-        logger.error('requeue web service error', error);
-      });
-  };
+  startTransfer = async () => inqueueUtils(this.props.task);
 
   // web service call to delete the call recording/transcript
-  deleteResources = async () => {
-    //  get instance of Flex manager
-    const mgr = Flex.Manager.getInstance();
-    const { taskSid, workflowSid, queueName, attributes } = this.props.task;
-    const { recordSid, transcriptSid } = attributes;
-    const data = {
-      mode: 'deleteRecordResources',
-      taskSid,
-      recordingSid: recordSid,
-      transcriptSid,
-      Token: mgr.user.token,
-      attributes,
-      workflowSid,
-      queueName,
-    };
-
-    return http
-      .postUrlEncoded(buildUrl(inqueueUtilsUri), data)
-      .then(() => {
-        logger.debug(data); // JSON data parsed by `response.json()` call
-      })
-      .catch((error) => {
-        logger.error('Delete resource web service error', error);
-      });
-  };
+  deleteResources = async () => inqueueUtils.deleteResource(this.props.task);
 
   startCall = async () => {
-    const mgr = Flex.Manager.getInstance();
-    const activityName = mgr.workerClient.activity.name;
+    const manager = Flex.Manager.getInstance();
+    const activityName = manager.workerClient.activity.name;
     if (activityName === 'Offline') {
       // eslint-disable-next-line no-alert
       alert('Change activity state from "Offline" to place call to contact');
@@ -309,29 +174,3 @@ class InQueueVoicemailComponent extends React.Component {
     );
   }
 }
-
-/*
- * Define app property mapping to Redux store state elements
- *
- * This maps application properties to Store state element values
- *  Syntax:
- *  <app_prop_name> : state[<namespace>].<Store_Identifier>.<State_Element>
- */
-const mapStateToProps = (state) => ({
-  vmCallButtonAccessibility: state['in-queue-redux'].InQueueMessaging.vmCallButtonAccessibility,
-  vmRecordButtonAccessibility: state['in-queue-redux'].InQueueMessaging.vmRecordButtonAccessibility,
-});
-
-/*
- *  Define mapping of local component methods to Redux Action methods for updating the store
- *
- *  Syntax:
- *  <Comp_Method> : bindActionCreators( Actions.<action_method>, dispatch)
- *
- */
-const mapDispatchToProps = (dispatch) => ({
-  vmCallButtonDisable: bindActionCreators(Actions.vmToggleCallButtonDisable, dispatch),
-  vmRecordButtonDisable: bindActionCreators(Actions.vmToggleRecordButtonDisable, dispatch),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(InQueueVoicemailComponent);
