+++ src/components/voicemail/VoicemailStyles.js
@@ -1,0 +1,57 @@
+export default {
+  itemWrapper: {
+    width: '100%',
+  },
+  itemBold: { fontWeight: 'bold' },
+  item: {
+    width: 150,
+  },
+  itemDetail: {
+    textAlign: 'right',
+    float: 'right',
+    marginRight: '5px',
+    marginTop: '3px',
+  },
+  cbButton: {
+    width: '100%',
+    marginBottom: '5px',
+    fontSize: '9pt',
+  },
+  audioWrapper: {
+    width: '100%',
+  },
+  transcriptWrapper: {
+    width: '100%',
+    marginBottom: 10,
+  },
+  h4Title: {
+    fontWeight: 'bold',
+  },
+  transcript: {
+    width: '90%',
+    marginRight: 10,
+    height: 50,
+    paddingLeft: 10,
+    paddingRight: 10,
+    paddingTop: 10,
+    paddingBottom: 10,
+    border: 'solid 1px #999999',
+  },
+  audio: {
+    width: '100%',
+    marginTop: 10,
+    marginBottom: 10,
+  },
+  textCenter: {
+    textAlign: 'center',
+    color: 'blue',
+  },
+  textAlert: {
+    color: 'red',
+    textAlign: 'center',
+  },
+  test: {
+    fontSize: '9pt',
+  },
+  info: { position: 'relative', top: '3px' },
+};
