+++ src/components/voicemail/index.js
@@ -1,0 +1,3 @@
+import VoicemailContainer from './VoicemailContainer';
+
+export default VoicemailContainer;
