+++ src/helpers/index.js
@@ -3,6 +3,4 @@ import * as http from './http';
 
 export { buildUrl } from './urlHelper';
 
-export { InQueueMessagingHelper } from './InQueueMessagingHelper';
-
 export { logger, http };
